Jdk 17 new (#161)
* [release] Prepare release 0.30.0

* Master 1.0 (#88)

* Start next iteration with 0.30.1 snapshot (#80)

* [release] Prepare release 0.30.0

* [skip ci] Start next iteration with 0.30.1-SNAPSHOT

* Updated version of secure token service (#82)

* Basic test for PopController (#58)

* STS-57. Bump to junit 5
* STS-57. Added E2E test package + very basic test for PopController
* STS-57. Test Jar added

* Feature/sts 59 e2e for key management (#61)

* STS-57. Bump to junit 5

* STS-57. Added E2E test package + very basic test for PopController

* STS-57. Test Jar added

* STS-59. More control for certificate creation time

* STS-59. More control for certificate creation time

* STS-59. Added basic e2e test for key rotation

* STS-65. Bump surefire

* Extracted API for keymanagement (#66)

* STS-64. Initial split of keymanagement into api and impl - moved
interfaces

* STS-64. Change packaging

* STS-64. Remove unnecessary dependency on impl of keymanagement

* STS-64. Move exceptions, drop unused dependency

* STS-64. Extract interfaces and replace concrete classes with interfaces

* STS-64. Using string interpolation for logging

* Feature/sts 60 rest controllers tests (#67)

* test for admin rest controller

* test for token exchange rest controller

* test for server info rest controller

* token exchange test

* added controllable clock

* grant password rest controller test

* tests without expexted same token

* authServer overrided

* test fixes

* json request and response from files

* token from file

* key retriever service from file

* Enable code coverage

* Enable code coverage

* Feature/sts 65 locking with schedlock (#70)

* STS-65. Initial split of keymanagement into api and impl - moved
interfaces
* STS-64. Change packaging
* STS-64. Remove unnecessary dependency on impl of keymanagement
* STS-64. Move exceptions, drop unused dependency
* STS-64. Extract interfaces and replace concrete classes with interfaces
* STS-64. Using string interpolation for logging
* STS-65. Using schedlock instead of customary locking solution
* STS-65. Added migration scripts
* STS-65. Added table alias
* STS-65. Restore test database and schema
* STS-65. Drop unused dependency
* STS-65. Fix PR comment

* Feature/sts 69 all databases migration (#75)

* STS-69. Key rotation clusterized environment test
* STS-69. Basic database migration tests
* STS-69. Another set of db-migration fixes
* STS-69. Perform a pop rotation for rdbms
* STS-69. Added mongo into test suite
* STS-69. Use annotation for test
* STS-69. Drop H2 test schema too
* STS-69. Added dirties context for cached keystore bean
* STS-69. Global dirties context
* STS-69. Destroy DB for crucial tests
* STS-69. Dirties context for concurrent key rotation
* STS-69. Add debug log
* STS-69. Add pool termination
* STS-69. Drop unnecessary log

* V2/documentation (#76)

* Add the explanation of francis made at 2019/09/12.

* update docu

* update docu

* update docu

* update docu

* update docu

* update docu

* Fixed typo in sts-usage.md

* changed link to develop-v2

* Feature/sts 14 integrate keystore mgmt (#81)

* STS-14. Drop STS-common and stub (// FIXME-cleanup) its usages to refactor

* STS-14. Fixed some integration stubs in persistence

* STS-14. Fixed cleanup stubs in persistence mapping

* STS-14. Key generation using keystore-management

* STS-14. Make e2e pass

* STS-14. Fixed key creation type restored code to drop as FIXME

* STS-14. Restore key rotation test

* STS-14. Drop unused imports

* STS-14. Return immutable entry

* STS-14. Add extra assertion

* STS-14. All tests green with basic key rotation, still SNAPSHOT
dependency used

* STS-14. Update to in-place key status update

* Resolve merges

* STS-14. Added one more test and fixed typo

* STS-14. Fixed typos

* STS-16. Fix JavaDoc (#84)

* STS-16. Fix JDK-8212233 (#85)

* STS-16. Pin OpenJdk8 as compiler (#86)

* [release] Prepare release 1.0.0

* Bump nimbus-jose-jwt from 7.2.1 to 7.9 (#83)

Bumps [nimbus-jose-jwt](https://bitbucket.org/connect2id/nimbus-jose-jwt) from 7.2.1 to 7.9.
- [Changelog](https://bitbucket.org/connect2id/nimbus-jose-jwt/src/master/CHANGELOG.txt)
- [Commits](https://bitbucket.org/connect2id/nimbus-jose-jwt/branches/compare/7.9..7.2.1)


* Hotfix/sts 17 improve logs (#92)

* STS-17 Improve logging: Exchanged token is invalid

* release scripts update

* fix release

* first try

* dependencies hibernate and sql

* fixes

* jackson

* finish update

* fix docker plugin versions

* fix generationStrategy in jpa-entities

* rewrite it-test

* fixed entityManagerError

* fix keycloak errors

* fix 403 error

* Updated Dependencies

* Update maven-failsafe-plugin version

* Remove duplicate dependency

* jdk-17-new

* Change port in test config

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

** changed Postgres

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

** changed Postgres

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

** changed Postgres

* Update pull-requests.yml

* Fixed tests by removing old catchException dependency
Added Junit4 maven surefire plugin

** changed Postgres

---------

+++ sts-persistence-jpa/src/main/java/de/adorsys/sts/persistence/jpa/entity/JpaSecret.java
@@ -1,9 +1,9 @@
 package de.adorsys.sts.persistence.jpa.entity;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 
 @Getter
 @Setter
@@ -13,7 +13,7 @@ public class JpaSecret {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     private String subject;
